requirements.txt
anthropic>=0.39.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
pyyaml>=6.0.1
schedule>=1.2.0
python-dotenv>=1.0.0
gitpython>=3.1.0
aiohttp>=3.9.0
lxml>=5.1.0
markdown>=3.5.0
Jinja2>=3.1.0
